Move message renderers to separate files
